:bomb: Change project file structure

+++ backend/api/models/answer.go
@@ -3,8 +3,8 @@ package models
 type Answer struct {
 	Id          string `db:"id" json:"id"`
 	Uid         string `db:"uid" json:"uid"`
-	TaskId      int64  `db:"task_id" json:"task"`
-	ConditionId int64  `db:"condition_id" json:"condition"`
-	AuthorId    int64  `db:"author_id" json:"author"`
+	TaskId      int    `db:"task_id" json:"task"`
+	ConditionId int    `db:"condition_id" json:"condition"`
+	AuthorId    int    `db:"author_id" json:"author"`
 	Reason      string `db:"reason" json:"reason"`
 }
