+++ backend/api/models/completionCode.go
@@ -2,5 +2,5 @@ package models
 
 type CompletionCode struct {
 	// CompletionCode : Completion code to confirm task completion for workers.
-	CompletionCode int64 `db:"completion_code" json:"completionCode"`
+	CompletionCode int `db:"completion_code" json:"completionCode"`
 }
